+++ ksconf/layer.py
@@ -10,6 +10,7 @@ from tempfile import NamedTemporaryFile
 from typing import Callable, Iterator, Match
 
 from ksconf.compat import Dict, List, Set, Tuple
+from ksconf.hooks import get_plugin_manager
 from ksconf.util.file import relwalk, secure_delete
 
 """
@@ -278,6 +279,11 @@ class LayerFile_Jinja2(LayerRenderedFile):
             undefined=StrictUndefined,
             loader=FileSystemLoader(self.layer.root),
             auto_reload=False)
+
+        # Call plugin for jinja environment tweaking
+        pm = get_plugin_manager()
+        pm.hook.modify_jinja_env(env=environment)
+
         environment.globals.update(self.layer.context.template_variables)
         return environment